src/main/webapp/app/entities/item-model/delete/item-model-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, fakeAsync, inject, TestBed, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ItemModelService } from '../service/item-model.service';

import { ItemModelDeleteDialogComponent } from './item-model-delete-dialog.component';

describe('Component Tests', () => {
  describe('ItemModel Management Delete Component', () => {
    let comp: ItemModelDeleteDialogComponent;
    let fixture: ComponentFixture<ItemModelDeleteDialogComponent>;
    let service: ItemModelService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [ItemModelDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(ItemModelDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(ItemModelDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(ItemModelService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
